+++ pywbemtools/pywbemcli/_common.py
@@ -25,6 +25,7 @@ import re
 try:
     from collections.abc import Sequence
 except ImportError:
+    # pylint: disable=deprecated-class
     from collections import Sequence
 
 import six